PLEASE READ - first commit
this is the current implementation of the detection architecture. It has some issues that I'm sorting out right now, I expect the final version completed by Wednesday. I modified the configuration to normalize the photos and apply padding where required to preserve the aspect ratio of the photos -> the output shape of the network is 256 x 256 currently, but can be easily modified. I wouldn't worry to much about this because I'm responsible for integrating these two networks into one working module