Updated Distortion Architecture
This is the final version of the distortion architecture. To receive binary masks from the model, see dist_inference.py. You just have to specify the directory for the image, the directory for the model, and the output paths for the maps and the binary masks. What's relevant to training the unet architecture is the output_masks_dir, that will hold all of the 4th channels you need for your image set.